Add genetrack_util.py to genetrack.py so that a .pyc file is not generated.
